requirements.txt
pillow==12.1.0
pyside6==6.10.1
requests==2.32.5